requirements.txt
pandas==2.1.4

python-dotenv==1.0.0

tqdm==4.66.1

pydantic==2.5.3